+++ Teknologi_Projekt/Teknologi_Projekt/Content/Content.mgcb
@@ -25,3 +25,10 @@
 /processorParam:TextureFormat=Color
 /build:button_rectangle_depth_flat.png;button_rectangle_depth_flat
 
+#begin UIFont.spritefont
+/importer:FontDescriptionImporter
+/processor:FontDescriptionProcessor
+/processorParam:PremultiplyAlpha=True
+/processorParam:TextureFormat=Compressed
+/build:UIFont.spritefont
+
